Update README.md
As a beginner I found it hard to get familiar with the provided code. I propose a simple description of the first steps on using your research to help other researchers to get something out of your research much faster and therefore increase your exposure in the community. Hope you like it.